requirements.txt
pyyaml
python-box